- Bug has been reworked to act like it did before. - Carnosaur has been reworked to act like it did before. - Boss now has 5 health and Carnosaur movement logic. - Housekeeping of now unused objects in the project.

+++ OwlMan/OwlMan/Scripts/Overlord.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 
 public class Overlord : Node
 {
-	public static float STANDARD_GRAVITY = 18f;
+	public static int STANDARD_GRAVITY = 18;
 
 	public static Vector2 LevelBoundsX;
 	public static Vector2 LevelBoundsY;
